Change BuyAsync

+++ src/IqOptionApiDotNet/IIqOptionApiDotNetClient.cs
@@ -29,7 +29,7 @@ namespace IqOptionApiDotNet
         Task<UsersAvailabilityResult> GetUsersAvailabilityAsync(string requestId, long[] userId);
 		Task<FinancialInformationResult> GetFinancialInformationAsync(string requestId, ActivePair pair);
         Task<bool> ChangeBalanceAsync(string requestId, long balanceId);
-        Task<BinaryOptionsResult> BuyAsync(string requestId, ActivePair pair, int size, OrderDirection direction, DateTimeOffset expiration);
+        Task<BinaryOptionsResult> BuyAsync(string requestId, ActivePair pair, Decimal size, OrderDirection direction, DateTimeOffset expiration);
         Task<CandleCollections> GetCandlesAsync(string requestId, ActivePair pair, TimeFrame tf, int count, DateTimeOffset to);
         Task<IObservable<CurrentCandle>> SubscribeRealtimeQuoteAsync(string requestId, ActivePair pair, TimeFrame tf);
